Enable data update
The application can now update its data when run, automatically. However,
we are using a serialized data format for storage i.e. RDS files, and this
will need to be reviewed. It's either the existing file is replaced on
updates, a database file is used or an API from the original data source
is called.