[Bug Fix]: Explicit loading for packages
In some of the functions, package loading was accomplished within
function scope. However, it appears that upon deploying the App remotely,
these packages were not parsed when loaded with a functional programming
paradigm (in this case with `lapply`). This was fixed notably in
`globals.R`, where we now use an individual call to `library` for each
required package.