Enable radio buttons for multi-country displays
In the case where more than one country is selected, we don't want to display both the cases and deaths in the same plot, as this makes it too cluttered. So, when more than one country is selected, the checkbox group for the variable is hidden and input is now governed by radio buttons, so the user selects only one of the two options.